7/31

+++ bn.h
@@ -5,7 +5,7 @@
 typedef struct bignum_st {
     uint32_t d[16];
     int top; //실제 사용중인 d 배열의 크기
-    int cb; //carry 또는 borrow
+    uint8_t cb; //carry 또는 borrow
 }BIGNUM;
 
 void initBignum(const uint32_t* num, const size_t inlen, BIGNUM* BN);
